Perform Probabilistic Quotient Normalization

Perform Probabilistic Quotient Normalization

normalize_data_pqn(x, pqn_reference = "median", pgn_reference_sample = NULL)

Arguments

x

matrix to normalize. row is variable and column is sample.

pqn_reference

normalization reference: "mean" for using the overall average of variables as reference or "median" (default) for using the overall median of variables as reference

pgn_reference_sample

vector of index to specify samples which average to use as reference

Value

Normalized expression dataset.

Details

First a total area normalization should be done before PQN is applied.

References

Dieterle, F., Ross, A., Schlotterbeck, G. & Senn, H. Probabilistic Quotient Normalization as Robust Method to Account for Dilution of Complex Biological Mixtures. Application in H1 NMR Metabonomics. Anal. Chem. 78, 4281-4290 (2006).

Author

E. Nevedomskaya

Rico Derks